Fix: Event handler binding issue
- 루트 컴포넌트의 이벤트 핸들러가 제대로 바인딩되지 않는 버그를 해결하였습니다.
- 자세한 수정 과정은 노션에 작성하였습니다
